Create a base list command
This command will be used to create multiple different listing methods
